display_items: Refactor into compiled module
display_items.h was previously included by every driver via
init_item() and destroy_items() function bodies in the ESP32 build
(and one more in the SDL build). Refactor into a proper compiled
module so the bodies are linked once.
