+++ CMakeLists.txt
@@ -29,6 +29,7 @@ endif()
 idf_component_register(SRCS
     "display_common.c"
     "display_driver.c"
+    "display_items.c"
     "5in65_acep_7c_display_driver.c"
     "font_data.c"
     "ili934x_display_driver.c"
